chore: made requested changes
