--- client/dashboard/app/utils/constants.ts
@@ -1,8 +1,0 @@
-enum States {
-    DRAFT= "draft",
-    ACCEPTED= "accepted",
-    REJECTED= "rejected",
-    PENDING="pending",
-}
-
-export default States;
